Refactor domain logic and data access
