+++ common/src/index.ts
@@ -4,5 +4,5 @@ export { Config } from '@/Config';
 export * from '@/constant';
 export * from '@/DataInterface';
 export * from '@/EventMap';
+export * from '@/QuasmComponent';
 export { ErrorCode, ErrorMap, extractMessage, QuasmError } from '@/QuasmError';
-export { QuasmComponent } from '@/Structure';
