+++ server/src/domain/game/character/Character.test.ts
