+++ server/src/domain/game/story/StoryChunk.ts
@@ -1,3 +1,9 @@
+export interface StoryChunkDetails {
+    title: string;
+    content: string;
+    imageUrl: string;
+}
+
 export class StoryChunk {
     constructor(
         readonly id: number = -1,
