+++ server/src/repositories/room/IRoomRepository.ts
@@ -1,25 +1,15 @@
-import { ChunkRange } from '@quasm/common';
 import { UUID } from 'crypto';
 
-import { Character, CharacterDetails } from '@/domain/game/Character';
-import {
-    ChatMessage,
-    ChatMessageDetails,
-    ChatParticipants
-} from '@/domain/game/ChatMessage';
-import { Room, RoomSettings } from '@/domain/game/Room';
-import { StoryChunk } from '@/domain/game/StoryChunk';
+import { CharacterDetails } from '@/domain/game/character/Character';
+import { Room } from '@/domain/game/room/Room';
+import { RoomSettingsDetails } from '@/domain/game/room/RoomSettings';
 
-/**
- * Main repository that manages instances of Rooms and their persistence
- */
 export interface IRoomRepository {
-    // fetchStoryChunks(roomID: string, range: ChunkRange): Promise<StoryChunk[]>;
     /**
      * Creates Room with single character (Game master) in it and returns created instance
      */
     createRoom(
-        roomDetails: RoomSettings,
+        roomDetails: RoomSettingsDetails,
         gameMasterDetails: CharacterDetails
     ): Promise<Room>;
 
@@ -36,57 +26,13 @@ export interface IRoomRepository {
     /**
      * Just persists soome room attributes changes
      */
-    updateRoom(roomID: UUID, roomSettings: RoomSettings): Promise<void>;
+    updateRoom(
+        roomID: UUID,
+        roomSettings: Partial<RoomSettingsDetails>
+    ): Promise<void>;
 
     /**
      * Destroys Room instance and deletes from persistent storage
      */
     deleteRoom(roomID: UUID): Promise<void>;
-
-    /**
-     * Instantiates character based on provided details adds it to Room and persists changes
-     */
-    addCharacter(
-        roomID: UUID,
-        characterDetails: CharacterDetails
-    ): Promise<Character>;
-
-    addMessage(chatMessageDetails: ChatMessageDetails): Promise<ChatMessage>;
-
-    /**
-     * Returns number of messages inside a given chat
-     */
-    fetchMessageCount(
-        chatParticipants: ChatParticipants,
-        roomID: UUID
-    ): Promise<number>;
-
-    fetchMessages(
-        chatParticipants: ChatParticipants,
-        range: ChunkRange
-    ): Promise<ChatMessage[]>;
-    /**
-     * Just persists some character attributes changes
-     */
-    updateCharacter(
-        id: UUID,
-        character: Partial<CharacterDetails>
-    ): Promise<void>;
-    /**
-     * Persists the change of the player's Submit
-     */
-    // setPlayerTurnSubmit(
-    //     id: UUID,
-    //     character: Partial<CharacterDetails>
-    // ): Promise<void>;
-
-    /**
-     * Adds StoryChunk to the specified Room
-     */
-    // addStoryChunk(roomID: UUID, storyChunk: StoryChunk): Promise<StoryChunk>;
-
-    /**
-     * Returns the requested part of the story - the StoryChunks specified in ChunkRange
-     */
-    // fetchStory(roomID: UUID, range: ChunkRange): Promise<StoryChunk[]>;
 }
